+++ src/gen_model.ml
@@ -369,6 +369,7 @@ let to_model (ast : A.ast) : M.model =
       | A.Pconst Coperations                   -> M.Moperations
       | A.Pconst Cmetadata                     -> M.Mmetadata
       | A.Pconst Ctotalvotingpower             -> M.Mtotalvotingpower
+      | A.Pconst Clevel                        -> M.Mlevel
       | A.Pconst c                             ->
         Format.eprintf "expr const unkown: %a@." A.pp_const c;
         assert false