+++ src/gen_why3.ml
@@ -2170,6 +2170,7 @@ let rec map_mterm m ctx (mt : M.mterm) : loc_term =
     | Mselfaddress          -> Tdoti(dl gs, dl (mk_id "selfaddress"))
     | Mchainid              -> Tchainid (dl gs)
     | Mmetadata             -> assert false
+    | Mlevel                -> assert false
 
     (* variables *)