+++ CHANGES.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - add new cryptographic hash functions: `sha3` and `keccak`
  - add voting instructions: `totalvotingpower` and `votingpower`
+ - add `level` constant
 
 ## Unreleased
 ### Added